apps/vectorisationApp/requirements.txt
fastapi==0.103.1
open_clip_torch==2.20.0
Pillow==10.0.1
torch==2.0.1
uvicorn==0.22.0
transformers==4.30.2
python-multipart==0.0.6